modules/data-modeling/requirements.txt
snowflake-connector-python==3.6.0
python-dotenv==1.0.0
requests==2.31.0
